OfferPage has been work upon
Load text from json-file of offers done.
Added button for applying for the project with halfway routing to login/application.
Stolendocument.js might not be useful but has been added just in case (might be useful for getting the information about the offer via url).

Kown issue: Applying for an offer while logged in clears the list of offers when adding the userId to the speciffic offer.

Co-Authored-By: rasje17 <rasje17@users.noreply.github.com>